Filter duplicate messages in queue to prevent spam and lag.
The exact same message twice in a queue results most likely from double clicking a link or spamming the bot (like if 100 people are klicking on a guild info link multiple times). There is no point in sending the exact same answer to the same person twice in a row. After the msg was sent, the user can - of course - ask again for the same info.